Created a link from the library page to each story page.
This was done by creating a "story" state in ui-router, with the story's id as reference. Then, the id
variable was added to the library controller, allowing to create the ui-sref (ui-router's link directive)
to the story's individual page. In order to conntect the ID to the story, ui-router's $stateParam was
injected to the story controller.